Project 3 - Jython
This is collage project I worked on using Jython. The program takes an image and edits it in various ways and copies the results to a blank canvas. The source images are planet.jpg and planet2.jpg. The output image is the_other_collage.jpg. I used a photocopy of my signature to sign the bottom right of the image.